Intégration de la lib_reg_lin comme une librairie visual studio ratachée à la solution. Cela permet de régler les problèmes d'inclusion. Première implémentation des options.

+++ replication portfolio/Portfolio.cpp
@@ -1,7 +1,7 @@
 #include "Portfolio.h"
 using namespace std;
 
-Portfolio::Portfolio(deque<float> nav, string ul, int maturity)
+Portfolio::Portfolio(deque<double> nav, string ul, int maturity)
 {
 	_nav = nav;
 	_ul = ul;
