+++ replication portfolio/greeks.cpp
@@ -1,16 +1,15 @@
 #include "greeks.h"
-#include "data_frame.h"
 
-const deque<float> vect_0;
-const deque<deque<float> > vect_00;
+const deque<double> vect_0;
+const deque<deque<double> > vect_00;
 
 // On approxime la dérivée par un taux d'accroissement
 // par convention, on prend la première valeur nulle, 
 // cela ne pose pas de problème dans la régression grâce au grand nombre de données
-deque<float> calculate_delta(deque<float> nav, string ul)
+deque<double> calculate_delta(deque<double> nav, string ul)
 {
-	deque<float> delta;
-	deque<float> value_ul = data.getnav(ul);
+	deque<double> delta;
+	deque<double> value_ul = data.getnav(ul);
 	delta[0] = 0; 
 	for (size_t i = 1; i < nav.size(); i++)
 	{
@@ -19,10 +18,10 @@ deque<float> calculate_delta(deque<float> nav, string ul)
 	return delta;
 }
 
-deque<float> calculate_gamma(deque<float> nav, string ul)
+deque<double> calculate_gamma(deque<double> nav, string ul)
 {
-	deque<float> gamma;
-	deque<float> delta = calculate_delta(nav, ul);
+	deque<double> gamma;
+	deque<double> delta = calculate_delta(nav, ul);
 	gamma[0] = 0;
 	gamma[1] = 0;
 	for (size_t i = 2; i < nav.size(); i++)
@@ -32,11 +31,11 @@ deque<float> calculate_gamma(deque<float> nav, string ul)
 	return gamma;
 }
 
-deque<float> calculate_vega(deque<float> nav, string ul)
+deque<double> calculate_vega(deque<double> nav, string ul)
 {
-	deque<float> vega;
+	deque<double> vega;
 	string vol_ul = "vol_"+ul;
-	deque<float> vol = data.getnav(vol_ul);
+	deque<double> vol = data.getnav(vol_ul);
 	vega[0] = 0;
 	for (size_t i = 1; i < nav.size(); i++)
 	{
@@ -45,10 +44,10 @@ deque<float> calculate_vega(deque<float> nav, string ul)
 	return vega;
 }
 
-deque<float> calculate_rho(deque<float> nav)
+deque<double> calculate_rho(deque<double> nav)
 {
-	deque<float> rho;
-	deque<float> value_r = data.getnav("LIBOR 3M USD");
+	deque<double> rho;
+	deque<double> value_r = data.getnav("LIBOR 3M USD");
 	rho[0] = 0;
 	for (size_t i = 1; i < nav.size(); i++)
 	{
@@ -57,9 +56,9 @@ deque<float> calculate_rho(deque<float> nav)
 	return rho;
 }
 
-deque<float> calculate_theta(deque<float> nav, int maturity)
+deque<double> calculate_theta(deque<double> nav, int maturity)
 {
-	deque<float> theta;
+	deque<double> theta;
 	theta[0] = 0;
 	for (int i = 1; i < maturity; i++)
 	{
