+++ replication portfolio/main.cpp
@@ -1,11 +1,12 @@
 #pragma once
 #include "portfolio.h"
 #include "data_frame.h"
-#include "..\\Librarie reg lin\dataanalysis.h"
+#include <dataanalysis.h>
 using namespace std;
 
 
 void main(){
-	float res = data._data[0][0];
+	double res = data._data[0][0];
 	cout << res;
 }
+
